[acpi] Prepare to build on the host.
This CL has no functional change. It adds a bunch of #ifdefs to the
existing ACPI code so that we can build it on the host, and a stub
implementation of acpi::Device that doesn't call into the DDK.

We also split Manager into FuchsiaManager (which basically just
implements all the async bits) and the HostManager (which stubs them
out, since they're not used on the host).

Bug: 98871
Change-Id: I4433a4a2df321f77bec8b2fca14fed45a050ac7b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/674813
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Simon Shields <simonshields@google.com>
Reviewed-by: Alex Legg <alexlegg@google.com>