[driver_tools] Match list-device DFv1/v2 behavior
Update the V2 code so that the old properties print human readable
strings which matches the V1 behavior.

Also update the V1 code to print:

 Key {} Value {}

which matches the V2 behavior.

Finally, update V1+V2 to pad the Key field to 30 characters,
so that most properties are aligned.

Change-Id: I3998da409e2511bc0fe987a14beb64fb5051ae30
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678856
Reviewed-by: Nick Eaton <nickeaton@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>