[nvme] Respect CAP.MQES for I/O queues.
The spec defines CAP.MQES as the maximum number of entries allowed in an
I/O queue. The NVME driver was written assuming that MQES would always
be greater than the number of queue entries that fit in one page.

This worked fine, until it didn't. This CL adds the appropriate logic to
restrict the size of I/O queues to either 1 page OR the value in MQES.

Bug: 100252
Test: `lsblk` on affected system now works.
Change-Id: I09e17c500882b280e179689116b0d8b4bf5abd89
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/679854
Reviewed-by: Christopher Anderson <cja@google.com>
Reviewed-by: Curtis Galloway <curtisgalloway@google.com>
Commit-Queue: Simon Shields <simonshields@google.com>