[fdf] Remove RemoteControl proxy from driver_tools
Remove the usage of RemoteControl in the driver_tools library. Instead,
functions are passed as arguments that provide proxies to FIDL services
only when needed. This is to avoid needing a connection to RemoteControl
in order to execute subcommands of the driver command. This enables the
driver command to have subcommands that do not need a connection to
RemoteControl.

Bug: 1000236
Change-Id: I8b7f217faa47438390140e1fe69eab7054c30118
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/679351
Reviewed-by: Casey Dahlin <sadmac@google.com>
Reviewed-by: Rich Kadel <richkadel@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Nick Eaton <nickeaton@google.com>