[fdf] Split out driver_index.fidl from fdf FIDL
The fdf FIDL is part of the SDK. The driver index FIDL is only used
internally and therefore not needed in the SDK. This change splits it
out into its own FIDL library and refactors all usages of it.

Bug: 97984
Tests: Existing unit and integration tests work. New tests are
not needed as there are no changes in behavior.

Change-Id: I76df16941821e308515b49617d757f92fd79a4d6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680948
Reviewed-by: Sarah Chan <spqchan@google.com>
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Jocelyn Dang <jocelyndang@google.com>
API-Review: Jocelyn Dang <jocelyndang@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>