[devfs] Make devfs_is_local check more specific
With the check as is, it is impossible to open a device which has an
outgoing directory as it is treated as a local node with the open
request not routed to the driver. This was happening due to
unintentionally inferring something about the device from a flag which
means something else. A new flag was introduced to more specifically
convey the real property the code was looking to identify.

Fixed: 101725
Change-Id: I9bb98f2158b1caf819ba017226477419b57e0c34
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686614
Fuchsia-Auto-Submit: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Ian Fisher <iafisher@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>