[driver_runtime] DFV2 integration - add runtime connector.
This is a temporary solution to allow the WLAN team to begin
proto-typing with DFv2. It will be removed once the proper solution
is implemented.

This allows the child driver to connect to parent runtime protocols.

The parent driver will offer the
fuchsia.driver.framework.RuntimeConnector protocol. Using this, the
child driver can then call driver::Connect() for the desired protocols.

The test mainly checks that the runtime connector works as expected.
Rather than doing an actual FIDL request on the runtime channel,
it just sends an empty request and reads an int back.

Bug: 87162

Change-Id: I2fc455f9b65275520b0e762b71bdf634da51859e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/660819
Commit-Queue: Jocelyn Dang <jocelyndang@google.com>
API-Review: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>