[fdf][compat] Fix logging on Remove errors
If we go to remove a node controller, but the node controller
has already been removed, the FIDL error will say that the
request has been cancelled.

In this case, we do not need to log an error. In this case, we
are trying to remove our child but it was already in the process
of being removed.

This fixes a flake where a driver calls device_async_remove
while it is shutting down, but its child is also in the process
of shutting down.

TEST: hid-driver-test-dfv2
Multiply: hid-driver-test-dfv2

Fixed: 101688
Fixed: 101311
Change-Id: I6914482cbe4c44d643528dcf193ea29d732c6a29
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/686973
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: David Gilhooley <dgilhooley@google.com>
Reviewed-by: Sarah Chan <spqchan@google.com>